# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VeriFactu SaaS API - Sistema de Facturación AEAT
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /taxpayers/{taxpayerId}/clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients for taxpayer
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Clients list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Products
...
  /taxpayers/{taxpayerId}/dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /taxpayers/{taxpayerId}/dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /taxpayers/{taxpayerId}/dashboard/aeat-status:
    get:
      operationId: getAeatStatus
      tags: [dashboard]
      summary: Get AEAT submission status overview
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AEAT status overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AeatStatusOverview"
  /taxpayers/{taxpayerId}/dashboard/vat-summary:
    get:
      operationId: getVatSummary
      tags: [dashboard]
      summary: Get VAT summary by period
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: VAT summary by period
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VatPeriodSummary"

  # Gestoria multi-client overview
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
      required:
        - email
        - password
        - name
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - createdAt

    # Organizations
...
    Client:
      type: object
      properties:
        id:
          type: integer
        taxpayerId:
          type: integer
        name:
          type: string
        nif:
          type: ["string", "null"]
        nifType:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        country:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - taxpayerId
        - name
        - country
        - isActive
        - createdAt
        - updatedAt
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        nif:
          type: ["string", "null"]
        nifType:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        country:
          type: string
      required:
        - name
        - country
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        nif:
          type: ["string", "null"]
        nifType:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        country:
          type: string

    # Product
...
    DashboardSummary:
      type: object
      properties:
        totalInvoices:
          type: integer
        totalEmitted:
          type: integer
        totalDrafts:
          type: integer
        totalCancelled:
          type: integer
        totalRevenue:
          type: number
        totalVat:
          type: number
        pendingSubmissions:
          type: integer
        acceptedSubmissions:
          type: integer
        rejectedSubmissions:
          type: integer
        revenueByMonth:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyRevenue"
        topClients:
          type: array
          items:
            $ref: "#/components/schemas/ClientRevenue"
      required:
        - totalInvoices
        - totalEmitted
        - totalDrafts
        - totalCancelled
        - totalRevenue
        - totalVat
        - pendingSubmissions
        - acceptedSubmissions
        - rejectedSubmissions
        - revenueByMonth
        - topClients
    MonthlyRevenue:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        invoiceCount:
          type: integer
      required:
        - month
        - revenue
        - invoiceCount
    ClientRevenue:
      type: object
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        revenue:
          type: number
        invoiceCount:
          type: integer
      required:
        - clientId
        - clientName
        - revenue
        - invoiceCount
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        entityId:
          type: ["integer", "null"]
        entityType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - description
        - createdAt
    AeatStatusOverview:
      type: object
      properties:
        pending:
          type: integer
        submitted:
          type: integer
        accepted:
          type: integer
        acceptedWithErrors:
          type: integer
        rejected:
          type: integer
        error:
          type: integer
        lastSubmission:
          type: ["string", "null"]
          format: date-time
      required:
        - pending
        - submitted
        - accepted
        - acceptedWithErrors
        - rejected
        - error
    VatPeriodSummary:
      type: object
      properties:
        period:
          type: string
        quarter:
          type: integer
        baseImponible:
          type: number
        vatRepercutido:
          type: number
        invoiceCount:
          type: integer
      required:
        - period
        - quarter
        - baseImponible
        - vatRepercutido
        - invoiceCount
...